# perf: skip type verify with -O