@geohot 2024-08-29
I think we need to think about this more. -O perf doesn't matter