@Qazalin 2024-08-29
maybe skip_check is a better place for this?
![image](https://github.com/user-attachments/assets/ac4863c3-76fe-4e18-a35f-934a1b9734a4)